requirements.txt
mcp[cli]>=1.6.0
httpx>=0.27.0
python-2captcha-solver>=1.0.5
pydub>=0.25.1
python-dotenv>=1.0.0
